requirements.txt
pandas
Pillow
torch
torchvision
torchsummary
